Moved HxLocation && HxTrigger* into different modules
